add new sendbird and zenzifa model

+++ src/model/doctor.go
@@ -7,7 +7,7 @@ import (
 )
 
 type Doctor struct {
-	ID                   uint64    `gorm:"primaryKey" json:"id"`
+	ID                   uint      `gorm:"primaryKey" json:"id"`
 	Name                 string    `gorm:"size:100;not null" json:"name"`
 	Photo                string    `gorm:"size:150;not null" json:"photo"`
 	Number               string    `gorm:"size:100" json:"number"`
